requirements.txt
beautifulsoup4==4.12.3
fastapi==0.112.0
gtfs-realtime-bindings==1.0.0
openai==1.41.1
pandas==2.2.2
pydantic==2.8.2
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.32.3
uvicorn==0.30.6